Plane metadata

Plane metadata for all plane tailnumbers found in the FAA aircraft registry. American Airways (AA) and Envoy Air (MQ) report fleet numbers rather than tail numbers so can't be matched.

Usage

planes

Format

A data frame with columns:

tailnum

Tail number.

year

Year manufactured.

type

Type of plane.

manufacturer, model

Manufacturer and model.

engines, seats

Number of engines and seats.

speed

Average cruising speed in mph.

engine

Type of engine.

Source

FAA Aircraft registry, https://www.faa.gov/licenses_certificates/aircraft_certification/aircraft_registry/releasable_aircraft_download

Examples

planes
#> # A tibble: 3,132 × 9
#>    tailnum  year type              manufacturer model engines seats speed engine
#>    <chr>   <int> <chr>             <chr>        <chr>   <int> <int> <int> <chr> 
#>  1 N10156   2004 Fixed wing multi… EMBRAER      EMB-…       2    55    NA Turbo…
#>  2 N102UW   1998 Fixed wing multi… AIRBUS INDU… A320…       2   182    NA Turbo…
#>  3 N103US   1999 Fixed wing multi… AIRBUS INDU… A320…       2   182    NA Turbo…
#>  4 N104UW   1999 Fixed wing multi… AIRBUS INDU… A320…       2   182    NA Turbo…
#>  5 N10575   2002 Fixed wing multi… EMBRAER      EMB-…       2    55    NA Turbo…
#>  6 N105UW   1999 Fixed wing multi… AIRBUS INDU… A320…       2   182    NA Turbo…
#>  7 N107US   1999 Fixed wing multi… AIRBUS INDU… A320…       2   182    NA Turbo…
#>  8 N108UW   1999 Fixed wing multi… AIRBUS INDU… A320…       2   182    NA Turbo…
#>  9 N109UW   1999 Fixed wing multi… AIRBUS INDU… A320…       2   182    NA Turbo…
#> 10 N110UW   1999 Fixed wing multi… AIRBUS INDU… A320…       2   182    NA Turbo…
#> # ℹ 3,122 more rows

library(dplyr, warn.conflicts = FALSE)
# Flights that don't have plane metadata
flights |>
  anti_join(planes, "tailnum")
#> # A tibble: 17,830 × 19
#>     year month   day dep_time sched_dep_time dep_delay arr_time sched_arr_time
#>    <dbl> <dbl> <dbl>    <int>          <int>     <dbl>    <int>          <int>
#>  1  2013     1     1      558            600        -2      742            730
#>  2  2013     1     1      603            610        -7      944            950
#>  3  2013     1     1      621            625        -4     1119           1105
#>  4  2013     1     1      705            700         5     1007            940
#>  5  2013     1     1      837            840        -3     1131           1115
#>  6  2013     1     1      904            910        -6     1221           1245
#>  7  2013     1     1     1010           1015        -5     1126           1135
#>  8  2013     1     1     1026           1030        -4     1324           1305
#>  9  2013     1     1     1108           1115        -7     1231           1245
#> 10  2013     1     1     1252           1255        -3     1410           1425
#> # ℹ 17,820 more rows
#> # ℹ 11 more variables: arr_delay <dbl>, carrier <chr>, flight <dbl>,
#> #   tailnum <chr>, origin <chr>, dest <chr>, air_time <dbl>, distance <dbl>,
#> #   hour <dbl>, minute <dbl>, time_hour <dttm>